requirements.txt
numpy
pandas
scikit-learn
scikit-surprise
fastapi
uvicorn